Add stubs for font changes that will be overriden by overlay
This CL defines stubs in TvSettings for id's that can be overridden in overlay so that we use downloadable fonts.

We define stubs for font_certs and for the fonts themselves. If a user defines these fonts in a runtime resource overlay, they can use these to override the sans-serif font in TvSettings using a style that uses downloadable fonts.

Figma mocks:
https://www.figma.com/design/13eLbD8z6GeBLT1xJRlQxl/DS-3?node-id=76030-24253&m=dev

Flag: NONE not needed
Bug: 355462390

Change-Id: I23ee5a08ca9090120a8710d4acdbb8020bcc594a