Refactor stubbed font changes
Instead of redefining brand_font and brand_font_medium in the overlay, we need to define the string attributes. This is because we need to be able to define these attributes with the app: mainspace in order for the font to show correctly, which we cannot do in the overlay because we would need to add extra dependencies to the RRO to achieve this.

Flag: NONE not needed
Bug: 331779785

Change-Id: Id654cd27aa4c5a0fddc6f2a13b976ae04d0b2450