Make dialog message scrollable
For the FullScreenDialog, wrap the message in a scroll view if it is too long. We only make this view focusable and scrollable if necessary. It only gets too long for current dialogs, if the user makes the text size larger, so this doesn't change anything for normal cases currently.

Flag: NONE not needed
Bug:384906263

Change-Id: Ib0e63ae1a9e486c4fead46f3b2db8ef699fa97f9