[TvSettings] fix PendingIntent conflict in RowBuilder actions
This issue is caused by RowBuilder when creating PendingIntents with the same ID.
When multiple intents share the same ID, subsequent RowBuilder instances reuse the first-created PendingIntent, resulting in incorrect behavior.
In this case, all `BluetoothActionActivity` instances receive only the Intent for the Rename action, regardless of the actual action triggered.

Bug: b/422009440
Test: Manual verification
Change-Id: I6e855de0d3bf257344c8440f5b76b22613b029cd