For Switch Preferences, add a custom state for Focusable When Disabled
To enable the behavior where a preference is focusable but the switch state is disabled, we declare a custom state called disabledButFocusable

FocusableWhenDisabledLinearLayout is needed so that the Preference layouts have the new state in their state arrays.

FocusableWhenDisabledPreference is needed to make sure the click event gets disabled when the custom state is enabled, and passes the disabledButFocusable to the LinearLayout, which passes it to the switch with duplicateParentState on the Switch

A preference with a switch would need to be changed to extend FocusableWhenDisabledPreference to use this behavior.

Example: https://hsv.googleplex.com/5947107847438336

Flag: NONE not needed
Bug: 331780378

Change-Id: Ia650eb21fd2cbb948203a0181e136c1ca7b357e4