Prevent Crash When Pressing Back Quickly After Changing Language
The crash is being caused by getChildFragmentManager(). During the language change procedure, the fragments are unattached, but the BACK operation runs on a delay.

If we try getChildfragmentManager(), then the an exception is thrown. Therefore, on the fragment, we add a guard to check if the host is null.

If the host is null, we do nothing.

Flag: NONE not needed bugfix
Bug: 427140024
Change-Id: Ibd2cfc2ce8d322c8113fbc50a649949fcad25433