Allow TvSettings to receive CSIS_SET_MEMBER_AVAILABLE intent
Currently, LE Unicast Audio only plays through one earbud on TV. There
are multiple reasons for this, but one of them is that TvSettings cannot
receive a necessary intent for progressing the CSIP pairing process.

This CL adds CSIS_SET_MEMBER_AVAILABLE to the intent filter for
BluetoothPairingRequest in TvSettings. This matches the corresponding
filter for BluetoothPairingRequest on Mobile.

This CL can be considered a followup to ag/30494568.

Bug: 433170421
Test: manual
Flag: android.media.tv.flags.enable_le_audio_unicast_ui
Change-Id: I9fe09f2d6037b204ce6a03190393e395eb030635