Disable surround sound options when volume control is variable
When the volume is set to variable control, surround sound is not
available because Android is unable to control volume of surround
sound formats. Therefore, options should be disabled and a
description should be shown to the user.

Also, show surround sound settings as Automatic and all surround
sound formats as disabled when surround sound is NONE.

Test: maunal
Flag: EXEMPT bugfix
Bug: 381025863
Change-Id: I116bc3c644bf836b609f405ff66aba65c140a477