Fix loadermanager error
The core issue is that getLoaderManager() in MainFragment is overridden b/c it is defined twice: for SuggestionControllerMixinCompat and AndroidX Fragment.

The AndroidX Fragment version is is NonNull, but the SuggestionControllermixinCompat version is not.

Kotlin version of MainFragment gives us a compiler warning that will be an error later on if we make it nullable. So we have to do something else.
Instead we re-implement SuggestionControllerMixinCompat from SettingsLib so that it will check if Fragment host is added before trying to fetch the loader.

This is basically the same as what is done for the java version, but more verbose.

Implementing this as java to make backporting easier if needed.

Flag: NONE not needed
Bug: 438389818
Change-Id: I787ac92dfee9423e64bd3eb494b7f4735fb53044