Refactor CaptionCustomFragment
- Declare needed constants at the top of the file
- Change opacity selection to not need constants in the code
- Reduce code duplication
- Remove unneeded view casts
- Mark all default values explicitly

Changing the captioning_opacity_selector_values array will not affect
existing devices since it is only ever an intermediary value and is not
saved as is in the secure settings.

Bug: 154360867
Bug: 154360018
Test: manual - factory reset and check that the default 'custom' caption
style matches the preview. Change all the caption options and make sure
they apply to the preview.
Flag: EXEMPT PURE_REFACTOR

Change-Id: Iea5f3c7708119c13552f3930d6f74e16bb509ee1