Update Thread Credentials Sharing Success behavior
In the design copy, when it's shared successfully, it
will show a pop-up and directly go back to the settings
page. But with the current implementation it will jump
to a success page with two buttons "Share again" and "OK".

The CL removes the success page and adds the pop-up.

Bug: 467554474
Test: Manual
Flag: NONE not needed
Change-Id: I26bccf1bdc76696e1d253edd194f8609d9470081